requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
MarkupSafe==2.1.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Session==0.5.0
google-generativeai==0.3.2
PyPDF2==3.0.1
textract==1.6.5
uuid==1.30
